Simplified alternation and serve loops operational
The implementation is much simplified (relative to the ThreadCSO implementation), because inports have poll() and outports have offer(), and their implementations complete only if a peer has already committed to the transfer (or in the case of offer on a buffer, has room for the transfer).

The poll cycle repeats after a (small) delay in this implementation.
This can and should be improved so that registered ports inform an alt when something changes.

Fairness is implemented straightforwardly

Update microCSO.scala