Started work on Resizeable
Fixed a problem with the timing of installation the installation of non-standard handlers: these can now be installed by overriding the whenStart method in  Application.scala (by which time the root glyph has been established).

The final parameter of an Interaction is a computation (informRoot) to be invoked when the GUI root has just been instructed to become visible. During its construction this RootGlyph (for it is usually such) has become known to its decendants.